+++ src/rust/proxy/momento/src/commands/mod.rs
@@ -1,0 +1,5 @@
+mod get;
+mod set;
+
+pub(crate) use get::*;
+pub(crate) use set::*;
