+++ src/rust/proxy/momento/src/main.rs
@@ -29,6 +29,11 @@ pub const MB: usize = 1024 * KB;
 const S: u64 = 1_000_000_000; // one second in nanoseconds
 const US: u64 = 1_000; // one microsecond in nanoseconds
 
+mod admin;
+mod commands;
+mod frontend;
+mod listener;
+
 // NOTES:
 //
 // This is a simple proxy which translates requests between memcache protocol
@@ -267,103 +272,12 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             );
             let tcp_listener =
                 TcpListener::from_std(tcp_listener).expect("could not convert to tokio listener");
-            listener(tcp_listener, client_builder, cache.cache_name()).await;
+            listener::listener(tcp_listener, client_builder, cache.cache_name()).await;
         });
     }
 
-    admin(log_drain, admin_listener).await;
+    admin::admin(log_drain, admin_listener).await;
     Ok(())
-}
-
-async fn listener(
-    listener: TcpListener,
-    client_builder: SimpleCacheClientBuilder,
-    cache_name: String,
-) {
-    // this acts as our listener thread and spawns tasks for each client
-    loop {
-        // accept a new client
-        if let Ok((socket, _)) = listener.accept().await {
-            TCP_ACCEPT.increment();
-
-            let client = client_builder.clone().build();
-            let cache_name = cache_name.clone();
-
-            // spawn a task for managing requests for the client
-            tokio::spawn(async move {
-                TCP_CONN_CURR.increment();
-                handle_proxy_client(socket, client, cache_name).await;
-
-                TCP_CLOSE.increment();
-                TCP_CONN_CURR.decrement();
-            });
-        }
-    }
-}
-
-async fn admin(mut log_drain: Box<dyn logger::Drain>, admin_listener: TcpListener) {
-    loop {
-        let _ = log_drain.flush();
-
-        // accept a new client
-        if let Ok(Ok((socket, _))) =
-            timeout(Duration::from_millis(1), admin_listener.accept()).await
-        {
-            TCP_CONN_CURR.increment();
-            TCP_ACCEPT.increment();
-            tokio::spawn(async move {
-                handle_admin_client(socket).await;
-                TCP_CLOSE.increment();
-                TCP_CONN_CURR.decrement();
-            });
-        };
-
-        let mut rusage = libc::rusage {
-            ru_utime: libc::timeval {
-                tv_sec: 0,
-                tv_usec: 0,
-            },
-            ru_stime: libc::timeval {
-                tv_sec: 0,
-                tv_usec: 0,
-            },
-            ru_maxrss: 0,
-            ru_ixrss: 0,
-            ru_idrss: 0,
-            ru_isrss: 0,
-            ru_minflt: 0,
-            ru_majflt: 0,
-            ru_nswap: 0,
-            ru_inblock: 0,
-            ru_oublock: 0,
-            ru_msgsnd: 0,
-            ru_msgrcv: 0,
-            ru_nsignals: 0,
-            ru_nvcsw: 0,
-            ru_nivcsw: 0,
-        };
-
-        if unsafe { libc::getrusage(libc::RUSAGE_SELF, &mut rusage) } == 0 {
-            RU_UTIME.set(rusage.ru_utime.tv_sec as u64 * S + rusage.ru_utime.tv_usec as u64 * US);
-            RU_STIME.set(rusage.ru_stime.tv_sec as u64 * S + rusage.ru_stime.tv_usec as u64 * US);
-            RU_MAXRSS.set(rusage.ru_maxrss * KB as i64);
-            RU_IXRSS.set(rusage.ru_ixrss * KB as i64);
-            RU_IDRSS.set(rusage.ru_idrss * KB as i64);
-            RU_ISRSS.set(rusage.ru_isrss * KB as i64);
-            RU_MINFLT.set(rusage.ru_minflt as u64);
-            RU_MAJFLT.set(rusage.ru_majflt as u64);
-            RU_NSWAP.set(rusage.ru_nswap as u64);
-            RU_INBLOCK.set(rusage.ru_inblock as u64);
-            RU_OUBLOCK.set(rusage.ru_oublock as u64);
-            RU_MSGSND.set(rusage.ru_msgsnd as u64);
-            RU_MSGRCV.set(rusage.ru_msgrcv as u64);
-            RU_NSIGNALS.set(rusage.ru_nsignals as u64);
-            RU_NVCSW.set(rusage.ru_nvcsw as u64);
-            RU_NIVCSW.set(rusage.ru_nivcsw as u64);
-        }
-
-        tokio::time::sleep(core::time::Duration::from_millis(100)).await;
-    }
 }
 
 async fn do_read(
@@ -403,476 +317,4 @@ async fn do_read(
             Err(e)
         }
     }
-}
-
-async fn stats_response(socket: &mut tokio::net::TcpStream) -> Result<(), Error> {
-    let mut data = Vec::new();
-    for metric in &metrics::common::metrics::metrics() {
-        let any = match metric.as_any() {
-            Some(any) => any,
-            None => {
-                continue;
-            }
-        };
-
-        // we need to filter some irrelvant metrics that
-        // are defined in crates we depend on
-        if metric.name().starts_with("add")
-            || metric.name().starts_with("append")
-            || metric.name().starts_with("cas")
-            || metric.name().starts_with("decr")
-            || metric.name().starts_with("delete")
-            || metric.name().starts_with("gets")
-            || metric.name().starts_with("incr")
-            || metric.name().starts_with("get_cardinality")
-            || metric.name().starts_with("ping")
-            || metric.name().starts_with("pipeline_depth")
-            || metric.name().starts_with("prepend")
-            || metric.name().starts_with("replace")
-            || metric.name().starts_with("request_latency")
-        {
-            continue;
-        }
-
-        if let Some(counter) = any.downcast_ref::<Counter>() {
-            data.push(format!("STAT {} {}\r\n", metric.name(), counter.value()));
-        } else if let Some(gauge) = any.downcast_ref::<Gauge>() {
-            data.push(format!("STAT {} {}\r\n", metric.name(), gauge.value()));
-        } else if let Some(heatmap) = any.downcast_ref::<Heatmap>() {
-            for (label, value) in PERCENTILES {
-                let percentile = heatmap.percentile(*value).unwrap_or(0);
-                data.push(format!(
-                    "STAT {}_{} {}\r\n",
-                    metric.name(),
-                    label,
-                    percentile
-                ));
-            }
-        }
-    }
-
-    data.sort();
-    ADMIN_RESPONSE_COMPOSE.increment();
-    for line in data {
-        socket.write_all(line.as_bytes()).await?;
-    }
-    socket.write_all(b"END\r\n").await
-}
-
-async fn handle_admin_client(mut socket: tokio::net::TcpStream) {
-    // initialize a buffer for incoming bytes from the client
-    let mut buf = Buffer::with_capacity(INITIAL_BUFFER_SIZE);
-
-    // initialize the request parser
-    let parser = AdminRequestParser::new();
-    loop {
-        if do_read(&mut socket, &mut buf).await.is_err() {
-            break;
-        }
-
-        ADMIN_REQUEST_PARSE.increment();
-
-        match parser.parse(buf.borrow()) {
-            Ok(request) => {
-                let consumed = request.consumed();
-                let request = request.into_inner();
-
-                match request {
-                    AdminRequest::Stats { .. } => {
-                        if stats_response(&mut socket).await.is_err() {
-                            break;
-                        }
-                    }
-                    _ => {
-                        debug!("unsupported command: {:?}", request);
-                    }
-                }
-                buf.consume(consumed);
-            }
-            Err(ParseError::Incomplete) => {}
-            Err(ParseError::Invalid) => {
-                // invalid request
-                let _ = socket.write_all(b"CLIENT_ERROR\r\n").await;
-                break;
-            }
-            Err(ParseError::Unknown) => {
-                // unknown command
-                let _ = socket.write_all(b"CLIENT_ERROR\r\n").await;
-                break;
-            }
-        }
-    }
-}
-
-async fn get(
-    client: &mut SimpleCacheClient,
-    cache_name: &str,
-    socket: &mut tokio::net::TcpStream,
-    keys: &[Box<[u8]>],
-) -> Result<(), Error> {
-    GET.increment();
-
-    // check if any of the keys are invalid before
-    // sending the requests to the backend
-    for key in keys.iter() {
-        if std::str::from_utf8(key).is_err() {
-            GET_EX.increment();
-
-            // invalid key
-            let _ = socket.write_all(b"ERROR\r\n").await;
-            return Err(Error::from(ErrorKind::InvalidInput));
-        }
-    }
-
-    let mut response_buf = Vec::new();
-
-    for key in keys.iter() {
-        BACKEND_REQUEST.increment();
-        GET_KEY.increment();
-
-        // we've already checked the keys, so we
-        // know this unwrap is safe
-        let key = std::str::from_utf8(key).unwrap();
-
-        match timeout(Duration::from_millis(200), client.get(&cache_name, key)).await {
-            Ok(Ok(response)) => {
-                match response.result {
-                    MomentoGetStatus::ERROR => {
-                        // we got some error from
-                        // the backend.
-                        BACKEND_EX.increment();
-
-                        GET_KEY_EX.increment();
-
-                        // TODO: what is the right
-                        // way to handle this?
-                        //
-                        // currently ignoring and
-                        // moving on to the next key
-                    }
-                    MomentoGetStatus::HIT => {
-                        GET_KEY_HIT.increment();
-
-                        let length = response.value.len();
-
-                        let item_header = format!("VALUE {} 0 {}\r\n", key, length);
-
-                        let response_len = 2 + item_header.len() + response.value.len();
-
-                        klog_get(key, response_len);
-
-                        response_buf.extend_from_slice(item_header.as_bytes());
-                        response_buf.extend_from_slice(&response.value);
-                        response_buf.extend_from_slice(b"\r\n");
-                    }
-                    MomentoGetStatus::MISS => {
-                        GET_KEY_MISS.increment();
-
-                        // we don't write anything for a miss
-
-                        klog_get(key, 0);
-                    }
-                }
-            }
-            Ok(Err(MomentoError::LimitExceeded(_))) => {
-                BACKEND_EX.increment();
-                BACKEND_EX_RATE_LIMITED.increment();
-                GET_KEY_EX.increment();
-            }
-            Ok(Err(e)) => {
-                // we got some error from the momento client
-                // log and incr stats and move on treating it
-                // as a miss
-                error!("error for get: {}", e);
-                GET_KEY_EX.increment();
-                BACKEND_EX.increment();
-            }
-            Err(_) => {
-                // we had a timeout, incr stats and move on
-                // treating it as a miss
-                GET_KEY_EX.increment();
-                BACKEND_EX.increment();
-                BACKEND_EX_TIMEOUT.increment();
-            }
-        }
-    }
-    response_buf.extend_from_slice(b"END\r\n");
-
-    SESSION_SEND.increment();
-    SESSION_SEND_BYTE.add(response_buf.len() as _);
-    TCP_SEND_BYTE.add(response_buf.len() as _);
-    if let Err(e) = socket.write_all(&response_buf).await {
-        SESSION_SEND_EX.increment();
-        return Err(e);
-    }
-    Ok(())
-}
-
-async fn set(
-    client: &mut SimpleCacheClient,
-    cache_name: &str,
-    socket: &mut tokio::net::TcpStream,
-    entry: MemcacheEntry,
-    noreply: bool,
-) -> Result<(), Error> {
-    SET.increment();
-
-    if let Ok(key) = std::str::from_utf8(&entry.key) {
-        let value = match entry.value {
-            Some(OwnedValue::Bytes(v)) => {
-                if let Ok(value) = std::str::from_utf8(&v) {
-                    value.to_owned()
-                } else {
-                    debug!("value is not valid utf8: {:?}", v);
-                    let _ = socket.write_all(b"ERROR\r\n").await;
-                    return Err(Error::from(ErrorKind::InvalidInput));
-                }
-            }
-            Some(OwnedValue::U64(v)) => {
-                format!("{}", v)
-            }
-            None => {
-                error!("empty values are not supported by momento");
-                SESSION_SEND.increment();
-                SESSION_SEND_BYTE.add(7);
-                TCP_SEND_BYTE.add(7);
-                if socket.write_all(b"ERROR\r\n").await.is_err() {
-                    SESSION_SEND_EX.increment();
-                }
-                return Err(Error::from(ErrorKind::InvalidInput));
-            }
-        };
-
-        BACKEND_REQUEST.increment();
-
-        let ttl = if let Some(ttl) = entry.ttl {
-            NonZeroU64::new(ttl.as_secs() as u64)
-        } else {
-            None
-        };
-
-        match timeout(
-            Duration::from_millis(200),
-            client.set(&cache_name, key, &value, ttl),
-        )
-        .await
-        {
-            Ok(Ok(result)) => {
-                match result.result {
-                    MomentoSetStatus::OK => {
-                        SET_STORED.increment();
-                        if noreply {
-                            klog_set(
-                                key,
-                                entry.flags,
-                                entry.ttl.map(|v| v.as_secs() as u32).unwrap_or(0),
-                                value.len(),
-                                5,
-                                0,
-                            );
-                        } else {
-                            klog_set(
-                                key,
-                                entry.flags,
-                                entry.ttl.map(|v| v.as_secs() as u32).unwrap_or(0),
-                                value.len(),
-                                5,
-                                8,
-                            );
-                            SESSION_SEND.increment();
-                            SESSION_SEND_BYTE.add(8);
-                            TCP_SEND_BYTE.add(8);
-                            if let Err(e) = socket.write_all(b"STORED\r\n").await {
-                                SESSION_SEND_EX.increment();
-                                // hangup if we can't send a response back
-                                return Err(e);
-                            }
-                        }
-                    }
-                    MomentoSetStatus::ERROR => {
-                        SET_NOT_STORED.increment();
-                        if noreply {
-                            klog_set(
-                                key,
-                                entry.flags,
-                                entry.ttl.map(|v| v.as_secs() as u32).unwrap_or(0),
-                                value.len(),
-                                9,
-                                0,
-                            );
-                        } else {
-                            klog_set(
-                                key,
-                                entry.flags,
-                                entry.ttl.map(|v| v.as_secs() as u32).unwrap_or(0),
-                                value.len(),
-                                9,
-                                12,
-                            );
-                            SESSION_SEND.increment();
-                            SESSION_SEND_BYTE.add(12);
-                            TCP_SEND_BYTE.add(12);
-                            // let client know this wasn't stored
-                            if let Err(e) = socket.write_all(b"NOT_STORED\r\n").await {
-                                SESSION_SEND_EX.increment();
-                                // hangup if we can't send a response back
-                                return Err(e);
-                            }
-                        }
-                    }
-                }
-            }
-            Ok(Err(MomentoError::LimitExceeded(_))) => {
-                BACKEND_EX.increment();
-                BACKEND_EX_RATE_LIMITED.increment();
-
-                SET_EX.increment();
-                SET_NOT_STORED.increment();
-                SESSION_SEND.increment();
-                SESSION_SEND_BYTE.add(12);
-                TCP_SEND_BYTE.add(12);
-
-                // let client know this wasn't stored
-                if let Err(e) = socket.write_all(b"NOT_STORED\r\n").await {
-                    SESSION_SEND_EX.increment();
-                    // hangup if we can't send a response back
-                    return Err(e);
-                }
-            }
-            Ok(Err(e)) => {
-                error!("error for set: {}", e);
-
-                BACKEND_EX.increment();
-                SET_EX.increment();
-                SET_NOT_STORED.increment();
-                SESSION_SEND.increment();
-                SESSION_SEND_BYTE.add(12);
-                TCP_SEND_BYTE.add(12);
-
-                // let client know this wasn't stored
-                if let Err(e) = socket.write_all(b"NOT_STORED\r\n").await {
-                    SESSION_SEND_EX.increment();
-                    // hangup if we can't send a response back
-                    return Err(e);
-                }
-            }
-            Err(_) => {
-                // timeout
-                BACKEND_EX.increment();
-                BACKEND_EX_TIMEOUT.increment();
-                SET_EX.increment();
-                SET_NOT_STORED.increment();
-                SESSION_SEND.increment();
-                SESSION_SEND_BYTE.add(12);
-                TCP_SEND_BYTE.add(12);
-
-                // let client know this wasn't stored
-                if let Err(e) = socket.write_all(b"NOT_STORED\r\n").await {
-                    SESSION_SEND_EX.increment();
-                    // hangup if we can't send a response back
-                    return Err(e);
-                }
-            }
-        }
-    } else {
-        SET_EX.increment();
-        // invalid key
-        SESSION_SEND.increment();
-        SESSION_SEND_BYTE.add(7);
-        TCP_SEND_BYTE.add(7);
-        if socket.write_all(b"ERROR\r\n").await.is_err() {
-            SESSION_SEND_EX.increment();
-        }
-        return Err(Error::from(ErrorKind::InvalidInput));
-    }
-    Ok(())
-}
-
-async fn handle_proxy_client(
-    mut socket: tokio::net::TcpStream,
-    mut client: SimpleCacheClient,
-    cache_name: String,
-) {
-    // initialize a buffer for incoming bytes from the client
-    let mut buf = Buffer::with_capacity(INITIAL_BUFFER_SIZE);
-
-    // initialize the request parser
-    let parser = MemcacheRequestParser::new(MAX_REQUEST_SIZE, TIME_TYPE);
-
-    // handle incoming data from the client
-    loop {
-        if do_read(&mut socket, &mut buf).await.is_err() {
-            break;
-        }
-
-        match parser.parse(buf.borrow()) {
-            Ok(request) => {
-                let consumed = request.consumed();
-                let request = request.into_inner();
-
-                match request {
-                    MemcacheRequest::Get { keys, .. } => {
-                        if get(&mut client, &cache_name, &mut socket, &keys)
-                            .await
-                            .is_err()
-                        {
-                            break;
-                        }
-                    }
-                    MemcacheRequest::Set { entry, noreply } => {
-                        if set(&mut client, &cache_name, &mut socket, entry, noreply)
-                            .await
-                            .is_err()
-                        {
-                            break;
-                        }
-                    }
-                    _ => {
-                        debug!("unsupported command: {}", request.command());
-                    }
-                }
-                buf.consume(consumed);
-            }
-            Err(ParseError::Incomplete) => {}
-            Err(ParseError::Invalid) => {
-                // invalid request
-                let _ = socket.write_all(b"CLIENT_ERROR\r\n").await;
-                break;
-            }
-            Err(ParseError::Unknown) => {
-                // unknown command
-                let _ = socket.write_all(b"CLIENT_ERROR\r\n").await;
-                break;
-            }
-        }
-    }
-}
-
-/// Logs GET or GETS
-fn klog_get(key: &str, response_len: usize) {
-    if response_len == 0 {
-        klog!("\"get {}\" 0 {}", key, response_len);
-    } else {
-        klog!("\"get {}\" 4 {}", key, response_len);
-    }
-}
-
-/// Logs SET, ADD, or REPLACE
-fn klog_set(
-    key: &str,
-    flags: u32,
-    ttl: u32,
-    value_len: usize,
-    result_code: usize,
-    response_len: usize,
-) {
-    klog!(
-        "\"set {} {} {} {}\" {} {}",
-        key,
-        flags,
-        ttl,
-        value_len,
-        result_code,
-        response_len
-    );
 }
